+++ Skittles/Engine.cpp
@@ -9,17 +9,17 @@ void Engine::init()
     return;
 }
 
-bool Engine::validate(int start, int end)
+bool Engine::validate(short int start, short int end)
 {
     return board.validate(start, end);
 }
 
-bool Engine::promote(int pieceNameVal)
+bool Engine::promote(short int pieceNameVal)
 {
     return board.promote(pieceNameVal);
 }
 
-int Engine::getSpecialMove()
+short int Engine::getSpecialMove()
 {
     return board.getSpecial();
 }
@@ -28,19 +28,19 @@ bool Engine::gameOver() {
     return board.gameOver();
 }
 
-int Engine::getWinner()
+short int Engine::getWinner()
 {
     return board.getWinner();
 }
 
-void Engine::goBack(int &start, int &end, int &special, int &promoPiece, int &capturedPiece, int &color){
+void Engine::goBack(short int &start, short int &end, short int &special, short int &promoPiece, short int &capturedPiece, short int &color){
     board.goBack(start, end, special, promoPiece, capturedPiece, color);
 }
 
-void Engine::goForward(int &start, int &end, int &special, int &promoPiece, int &capturedPiece, int &color){
+void Engine::goForward(short int &start, short int &end, short int &special, short int &promoPiece, short int &capturedPiece, short int &color){
     board.goForward(start, end, special, promoPiece, capturedPiece, color);
 }
 
-void Engine::moveStats(int &pieceMoved, int &color, bool &capture, bool &check, bool &checkmate){
+void Engine::moveStats(short int &pieceMoved, short int &color, bool &capture, bool &check, bool &checkmate){
     board.moveStats(pieceMoved, color, capture, check, checkmate);
 }
