+++ Skittles/Move.h
@@ -9,14 +9,14 @@ private:
     Move *next = nullptr;
 
 public:
-    Move(int startPos, int endPos, int passantArray[2], int passantColor, int special, bool capture, int promo, ChessPiece caturedPiece, bool mate, int moved);
-    int start;                      // where the piece started this move
-    int end;                        // where the piece ended this move
-    int enPassant[2] = {};          // enpassant information
-    int enPassantColor;             // enpassant information
-    int specialMove;                // special move that was active after this move completed
+    Move(short int startPos, short int endPos, short int passantArray[2], short int passantColor, short int special, bool capture, short int promo, ChessPiece caturedPiece, bool mate, short int moved);
+    short int start;                      // where the piece started this move
+    short int end;                        // where the piece ended this move
+    short int enPassant[2] = {};          // enpassant information
+    short int enPassantColor;             // enpassant information
+    short int specialMove;                // special move that was active after this move completed
     bool captured;                  // whether the move captured anotherpiece or not
-    int promoTo;                    // piece type piece was promoted to
+    short int promoTo;                    // piece type piece was promoted to
     ChessPiece piece;               // piece that was captured
     bool mate;                      // whether this move caused mate
     bool moved;                     // whether the piece that was moved had moved before
